# openapi/aisi.v1.yaml
openapi: 3.0.3
info:
  title: AISI / Dakwah Depok API
  version: 1.0.0
  description: |
    Kontrak SSOT API AISI (Go).
    Envelope sukses: `{ success: true, data, message?, pagination? }`
    Envelope error: `{ success: false, message }`
servers:
  - url: /api/v1
paths:
...
  /schools/{id}/groups:
    post:
      operationId: createSchoolGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '201':
          description: Created
...
  /groups:
    get:
      operationId: listGroups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      operationId: createGroup
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /groups/{id}:
    get:
      operationId: getGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
    put:
      operationId: updateGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
    delete:
      operationId: deleteGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
  /groups/{id}/members:
    get:
      operationId: listGroupMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
    post:
      operationId: addGroupMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '201':
          description: Created
  /groups/{id}/members/{userId}:
    get:
      operationId: getGroupMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    put:
      operationId: updateGroupMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    delete:
      operationId: removeGroupMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /analytics/group/{id}:
    get:
      operationId: analyticsGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
...
  /mutabaah/group/{groupId}:
    get:
      operationId: getGroupMutabaah
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...